Relocate KillException and QuitException to fix dependencies; fix handling of them in one case
