+++ truffle/com.oracle.truffle.api/src/com/oracle/truffle/api/instrument/QuitException.java
@@ -29,6 +29,7 @@ import com.oracle.truffle.api.nodes.ControlFlowException;
 /**
  * Controls breaking out of all executions and ending Truffle execution.
  */
+@Deprecated
 public final class QuitException extends ControlFlowException {
 
     private static final long serialVersionUID = -4301115629772778413L;
