+++ truffle/com.oracle.truffle.tools.debug.shell/src/com/oracle/truffle/tools/debug/shell/client/SimpleREPLClient.java
@@ -24,12 +24,6 @@
  */
 package com.oracle.truffle.tools.debug.shell.client;
 
-import com.oracle.truffle.api.instrument.QuitException;
-import com.oracle.truffle.api.source.Source;
-import com.oracle.truffle.tools.debug.shell.REPLClient;
-import com.oracle.truffle.tools.debug.shell.REPLMessage;
-import com.oracle.truffle.tools.debug.shell.server.REPLServer;
-
 import java.io.IOException;
 import java.io.PrintStream;
 import java.util.ArrayList;
@@ -43,6 +37,12 @@ import java.util.Map;
 import java.util.TreeSet;
 
 import jline.console.ConsoleReader;
+
+import com.oracle.truffle.api.QuitException;
+import com.oracle.truffle.api.source.Source;
+import com.oracle.truffle.tools.debug.shell.REPLClient;
+import com.oracle.truffle.tools.debug.shell.REPLMessage;
+import com.oracle.truffle.tools.debug.shell.server.REPLServer;
 
 /**
  * A very simple line-oriented, language-agnostic debugging client shell: the first step toward a
